pyproject.toml
[project]
name = "auth-service"
version = "0.1.0"
description = "Add your description here"
dependencies = [
    "fastapi>=0.120.1",
    "uvicorn>=0.38.0",
    "sqlalchemy[asyncio]>=2.0.31",
    "asyncpg>=0.29.0",
    "python-jose[cryptography]>=3.3.0",
    "pydantic-settings>=2.4.0",
    "passlib[bcrypt]>=1.7.4",
]
# readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/auth_service"]
